Changing the rate_pay to a float instead of in intager.

+++ Pay_Rate.py
@@ -3,8 +3,8 @@ rate = input('Enter Rate: ')
 
 try:
     hours_pay = int(hours)
-    rate_pay = int(rate)
-    
+    rate_pay = float(rate)
+
     if int(hours) > 40:
         normal_pay = 40 * rate_pay
         overtime_pay = (hours_pay - 40) * (rate_pay * 1.5)
